Issue #OS-121 MediaTypeConfiguration: to get configuration based on given media and direction
