+++ java/middleware-commons/src/main/java/io/opensaber/registry/transform/MediaTypeConfiguration.java
@@ -1,0 +1,24 @@
+package io.opensaber.registry.transform;
+
+import org.springframework.http.MediaType;
+import org.springframework.stereotype.Component;
+
+import io.opensaber.registry.middleware.util.Constants.Direction;
+
+@Component
+public class MediaTypeConfiguration {
+	
+	public static final String MEDIATYPE_APPLICATION_JSONLD = "application/ld+json";
+
+	public Configuration getConfiguration(String mediaType, Direction direction){
+		
+		if(mediaType.equalsIgnoreCase(MediaType.APPLICATION_JSON_VALUE) && direction == Direction.OUT){
+			return Configuration.LD2JSON;
+		}else if(mediaType.equalsIgnoreCase(MEDIATYPE_APPLICATION_JSONLD)){
+			return Configuration.LD2LD;
+		}else if(mediaType.equalsIgnoreCase(MediaType.APPLICATION_JSON_VALUE) && direction == Direction.IN){
+			return Configuration.JSON2LD;
+		}		
+		return null;
+	}
+}
