+++ java/middleware-commons/src/main/java/io/opensaber/registry/transform/Transformer.java
@@ -1,13 +1,9 @@
 package io.opensaber.registry.transform;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.MediaType;
-
-import io.opensaber.registry.middleware.util.Constants.Direction;
 
 public class Transformer {
 
-	public static final String MEDIATYPE_APPLICATION_JSONLD = "application/ld+json";
 	private static final String EXCEPTION_MESSAGE = "Media type not suppoted";
 
 	@Autowired
@@ -20,7 +16,7 @@ public class Transformer {
 	private Ld2LdTransformer Ld2LdTransformer;
 	
 	@Autowired
-	private LD2RDFTransformer ld2RDF;
+	private LD2RDFTransformer ld2RDFTransformer;
 	
 	public ITransformer<Object> getInstance(Configuration config) throws TransformationException {
 		ITransformer<Object> transformer = null;
@@ -32,24 +28,12 @@ public class Transformer {
 		}else if(config == Configuration.LD2LD){
 			transformer = Ld2LdTransformer;
 		}else if(config == Configuration.LD2RDF){
-			transformer = ld2RDF;				
+			transformer = ld2RDFTransformer;				
 		}else{
 			throw new TransformationException(EXCEPTION_MESSAGE, ErrorCode.UNSUPPOTERTED_TYPE);
 
 		}			
 		return transformer;
-	}
-	
-	public Configuration getConfiguration(String mediaType, Direction direction){
-		
-		if(mediaType.equalsIgnoreCase(MediaType.APPLICATION_JSON_VALUE) && direction == Direction.OUT){
-			return Configuration.LD2JSON;
-		}else if(mediaType.equalsIgnoreCase(MEDIATYPE_APPLICATION_JSONLD)){
-			return Configuration.LD2LD;
-		}else if(mediaType.equalsIgnoreCase(MediaType.APPLICATION_JSON_VALUE) && direction == Direction.IN){
-			return Configuration.JSON2LD;
-		}		
-		return null;
 	}
 
 }
