+++ java/registry/src/main/java/io/opensaber/registry/app/OpenSaberApplication.java
@@ -5,7 +5,7 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.ComponentScan;
 
 @SpringBootApplication
-@ComponentScan(basePackages = { "io.opensaber.registry", "io.opensaber.pojos"})
+@ComponentScan(basePackages = { "io.opensaber.registry", "io.opensaber.pojos", "io.opensaber.registry.transform"})
 public class OpenSaberApplication {
 	public static void main(String[] args) {
 		SpringApplication.run(OpenSaberApplication.class, args);
